requirements.txt
cnocr[ort-cpu]
flask
keyboard
toml
psycopg2-binary
zhipuai
pyside6
pyinstaller